requirements.txt
pandas
scikit-learn
imblearn
joblib
numpy
scipy
seaborn
fastapi
requests
pyyaml
colorama
tqdm
optuna
xgboost
fastapi
uvicorn
streamlit
streamlit-autorefresh